Fix postProcessors having value semantics when unset
Currently, if postProcessors was empty, it would be set to a new
pointer value. But that would only apply for the current Configurable,
so if you passed a Configurable by value to a function and then
that function added a postProcessor, the caller wouldn't see that.

Configurable properties in general have a confusing mix of pointer
and value semantics (thanks go...), this fixes one part of it but in
followup cls I'll try to make it consistent.

Bug: 362579941
Test: m nothing --no-skip-soong-tests
Change-Id: Ib33497768be3af45233cf1a682adffb28ee64bfd
Signed-off-by: Dmitrii <bankersenator@gmail.com>